+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 7.2.1 - 2021-05-24
+### Changed
+- `hive-hooks` component to convert Storage Description 'path' parameter.
+- `hive-hooks` component to convert table parameter 'avro.schema.url'.
+
 ## 7.2.0 - 2021-03-16
 ### Added
 - Added `hive-hooks` component for generic path and table location modification.